Module benchmark::bin_utils

Functions

create_ac_clients
Creat a vector of AdmissionControlClient and connect them to validators.

create_benchmarker_from_opt

gen_and_mint_accounts
Generate a group of new accounts, and mint them using Benchmarker before returning them.

linear_search_max_throughput
Search the maximum throughput by increasing submission rate from lower_bound to upper_bound. Search may end earilier if commit ratio < COMMIT_RATIO_THRESHOLD. Return the max observed TXN throughput, along with the request throughput.

measure_throughput
Play given TXNs with Benchmarker for several epochs and measure throughput, e.g., the average committed txns per second. Since time is counted from submission until all TXNs are committed, this measurement is in a sense the user-side throughput. Each epoch plays the given TXN pattern sequence repeatedly for several rounds. Return list of summaries for all epochs.

try_start_metrics_server
Benchmarker is not a long-lived job, so starting a server and expecting it to be polled continuously is not ideal. Directly pushing metrics when benchmarker is running can be achieved by using Pushgateway.