Module benchmark::grpc_helpers

Functions

divide_items
Divide generic items into a vector of chunks of nearly equal size.

get_account_states
Request a bunch of accounts' states, including sequence numbers and status from validator. Ignore any failure, during either requesting or processing, and continue for next account. Return the mapping from address to (sequence number, account status) tuple for all successfully requested accounts.

submit_and_wait_requests
Send requests using specified rate to AC async, wait for and check the responses (currently only for write requests). Return only the responses of accepted TXNs. Ignore but count both gRPC-failed submissions and AC-rejected requests.

sync_account_sequence_number
For each sender account, synchronize its persisted sequence number from validator. When this sync sequence number equals the account's local sequence number, all its transactions are committed. Timeout if such condition is never met for all senders. Return sender accounts' most recent persisted sequence numbers.