Module benchmark::load_generator

Structs

PairwiseTransferTxnGenerator
Pre-generate TXNs of pairwise transfers between accounts, including self to self transfer. For example, given account (A1, A2, A3, ..., AN), this method returns a vector of TXNs like (A1->A1, A1->A2, ..., A1->AN, A2->A1, A2->A2, ... A2->AN, ..., AN->A(N-1)).

RingTransferTxnGenerator
-------------------------------------------------------------------------------- /// Two LoadGenerator examples: circular transfer TXNs and pairwise transfer TXNs. /// -------------------------------------------------------------------------------- /// Pre-generate a ring/circle of transfer TXNs. For example, given account (A1, A2, A3, ..., AN), this method returns a vector of TXNs like (A1->A2, A2->A3, A3->A4, ..., AN->A1).

Enums

Request
This enum unifies both write and read requests. Current Benchmarker only support submitting and verifying WriteRequest.

Constants

TXN_EXPIRATION

Traits

LoadGenerator
This interface specifies the requirements to generate customized read/write requests that can be put into Benchmarker for playing. Required methods are expected to be called in following specified order:

Functions

gen_accounts_from_wallet
Create a number of accounts without keypair from a wallet.

gen_get_txn_by_sequnece_number_request
Generate a GetAccountTransactionBySequenceNumber request.

gen_mint_txn_requests
For each account, generate a mint TXN with the valid faucet account.

gen_repeated_requests
Generate repeated requests from a type that implements LoadGenerator.