Trait benchmark::load_generator::LoadGenerator

pub trait LoadGenerator {
    fn gen_accounts(&mut self, num_accounts: u64) -> Vec<AccountData>;

    fn gen_setup_requests(
        &self, 
        faucet_account: &mut AccountData, 
        accounts: &mut [AccountData]
    ) -> Vec<Request>;

    fn gen_requests(&self, accounts: &mut [AccountData]) -> Vec<Request>;
}

This interface specifies the requirements to generate customized read/write requests that can be put into Benchmarker for playing. Required methods are expected to be called in following specified order:

Required methods

fn gen_accounts(&mut self, num_accounts: u64) -> Vec<AccountData>

Generate arbitrary number of accounts.

fn gen_setup_requests(
&self,
faucet_account: &mut AccountData,
accounts: &mut [AccountData]
) -> Vec<Request>

Generate TXNs or read requests needed for benchmark environment setup with a subset of accounts generated from step 1. For example, minting accounts. It is OK to return empty vector.

fn gen_requests(&self, accounts: &mut [AccountData]) -> Vec<Request>

Generate arbitrary read/write requests from subset of accounts from step 1.

Loading content...